Add SelectFactory and simplify Select generic
The `SelectFactory` class provides a way to easily re-use a `Select` instance,
in scenarios where the statement structure remains the same, but the targets
or values change. This targets common control loop scenarios, where the select
statement is used to handle signaling.

To facilitate `SelectFactory` and other implementations that might encapsulate
the `Select` class, the automatic (type) unwrapping of `PromiseLike` has
been moved from the constructor, to a dedicated static method, `promises`.
This means that the generic of the `Select` is now a simple array or tuple of
cases.